Split SIMD tests into individual unit test files
