+++ simd/unit_tests/CMakeLists.txt
@@ -1,3 +1,5 @@
+KOKKOS_INCLUDE_DIRECTORIES(${KOKKOS_SOURCE_DIR}/simd/unit_tests/include)
+
 KOKKOS_ADD_EXECUTABLE_AND_TEST(
   UnitTest_SIMD
   SOURCES
